#2 fix relationship 개선

+++ src/main/java/com/pettory/pettory/jointshopping/command/domain/service/JointShoppingGroupDomainService.java
@@ -4,19 +4,16 @@ import com.pettory.pettory.exception.BadJoinException;
 import com.pettory.pettory.exception.NotFoundException;
 import com.pettory.pettory.jointshopping.command.application.dto.JointShoppingDeliveryInfoRequest;
 import com.pettory.pettory.jointshopping.command.application.dto.JointShoppingGroupRequest;
-import com.pettory.pettory.jointshopping.command.application.dto.JointShoppingGroupUserRequest;
+import com.pettory.pettory.jointshopping.command.domain.aggregate.JointShoppingCategory;
 import com.pettory.pettory.jointshopping.command.domain.aggregate.JointShoppingGroup;
-import com.pettory.pettory.jointshopping.command.domain.aggregate.JointShoppingGroupUser;
-import com.pettory.pettory.jointshopping.command.domain.aggregate.JointShoppingParticipationUser;
 import com.pettory.pettory.jointshopping.command.domain.repository.JointShoppingGroupRepository;
 import com.pettory.pettory.jointshopping.command.mapper.JointShoppingGroupMapper;
 import com.pettory.pettory.jointshopping.util.FileUploadUtils;
+import com.pettory.pettory.user.command.domain.aggregate.User;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
-
-import java.util.List;
 
 @Service
 @RequiredArgsConstructor
@@ -28,17 +25,17 @@ public class JointShoppingGroupDomainService {
     private final JointShoppingGroupRepository jointShoppingGroupRepository;
 
     /* 도메인 객체를 생성하는 로직 */
-    public JointShoppingGroup createGroup(JointShoppingGroupRequest groupRequest, MultipartFile productImg) {
+    public JointShoppingGroup createGroup(JointShoppingGroupRequest groupRequest, MultipartFile productImg, User user, JointShoppingCategory JointShoppingCategory) {
 
         /* 전달 된 파일을 서버의 지정 경로에 저장
-        * 파일이 없으면 저장 안함 */
+         * 파일이 없으면 저장 안함 */
         String replaceFileName = null;
-        if(productImg != null) {
+        if (productImg != null) {
             replaceFileName = IMAGE_DIR + FileUploadUtils.saveFile(IMAGE_DIR, productImg);
         }
 
         /* dto to entity */
-        JointShoppingGroup newJointShoppingGroup = JointShoppingGroupMapper.toEntity(groupRequest, replaceFileName);
+        JointShoppingGroup newJointShoppingGroup = JointShoppingGroupMapper.toEntity(groupRequest, replaceFileName, user, JointShoppingCategory);
 
         return newJointShoppingGroup;
     }
@@ -49,13 +46,17 @@ public class JointShoppingGroupDomainService {
     }
 
     /* 도메인 객체를 수정하는 로직 */
-    public JointShoppingGroup updateGroup(Long jointShoppingGroupNum, @Valid JointShoppingGroupRequest groupRequest, MultipartFile productImg) {
+    public JointShoppingGroup updateGroup(Long jointShoppingGroupNum, @Valid JointShoppingGroupRequest groupRequest, MultipartFile productImg, User user, JointShoppingCategory jointShoppingCategory) {
 
         JointShoppingGroup jointShoppingGroup = jointShoppingGroupRepository.findById(jointShoppingGroupNum)
                 .orElseThrow(() -> new NotFoundException("해당 번호에 맞는 모임이 없습니다. code : " + jointShoppingGroupNum));
 
+        if(!user.getUserId().equals(jointShoppingGroup.getUser().getUserId()) ){
+            throw new BadJoinException("방장이 아니여서 수정하실 수 없습니다.");
+        }
+
         /* 이미지 수정이 필요할 경우 새로운 이미지 저장 후 기존 이미지 삭제 */
-        if(productImg != null) {
+        if (productImg != null) {
             String replaceFileName = FileUploadUtils.saveFile(IMAGE_DIR, productImg);
             FileUploadUtils.deleteFile(IMAGE_DIR, jointShoppingGroup.getJointShoppingProductsFileDirectory().replace(IMAGE_DIR, ""));
             jointShoppingGroup.changejointShoppingProductsFileDirectory(IMAGE_DIR + replaceFileName);
@@ -69,49 +70,39 @@ public class JointShoppingGroupDomainService {
                 groupRequest.getJointShoppingCost(),
                 groupRequest.getJointShoppingGroupMaximumCount(),
                 groupRequest.getJointShoppingParticipationMaximumCount(),
-                groupRequest.getHostCourierCode(),
-                groupRequest.getHostInvoiceNum(),
-                groupRequest.getJointShoppingCategoryNum(),
-                groupRequest.getUserId()
+                jointShoppingCategory,
+                user
         );
 
         return jointShoppingGroup;
     }
 
-    /* 방 상태를 체크하고 리턴하는 로직 */
-    public void checkGroupState(Long jointShoppingGroupNum){
-
-        JointShoppingGroup jointShoppingGroup = jointShoppingGroupRepository.findById(jointShoppingGroupNum)
+    /* 모임 검색 */
+    public JointShoppingGroup findGroup(Long jointShoppingGroupNum) {
+        return jointShoppingGroupRepository.findById(jointShoppingGroupNum)
                 .orElseThrow(() -> new NotFoundException("해당 번호에 맞는 모임이 없습니다. code : " + jointShoppingGroupNum));
 
+    }
+
+    /* 방 상태를 체크하고 리턴하는 로직 */
+    public void checkGroupState(JointShoppingGroup jointShoppingGroup) {
         String state = jointShoppingGroup.getJointShoppingGroupState().toString();
-        if(!state.equals("APPLICATION")){
+        if (!state.equals("APPLICATION")) {
             throw new BadJoinException("방에 참가하실 수 없습니다.");
         }
     }
 
     /* 도메인 객체를 삭제하는 로직 */
-    public void deleteGroup(Long jointShoppingGroupNum) {
+    public void deleteGroup(Long jointShoppingGroupNum, User user) {
+        JointShoppingGroup jointShoppingGroup = jointShoppingGroupRepository.findById(jointShoppingGroupNum)
+                .orElseThrow(() -> new NotFoundException("해당 번호에 맞는 모임이 없습니다. code : " + jointShoppingGroupNum));
+
+        if(!user.getUserId().equals(jointShoppingGroup.getUser().getUserId()) ){
+            throw new BadJoinException("방장이 아니여서 삭제하실 수 없습니다.");
+        }
+
         /* soft delete 될 수 있도록 entity에 설정함 */
         jointShoppingGroupRepository.deleteById(jointShoppingGroupNum);
-    }
-
-    /* 최대 사용자 수를 반환하는 로직 */
-    public Integer findGroupMaximumCount(Long jointShoppingGroupNum) {
-
-        JointShoppingGroup jointShoppingGroup = jointShoppingGroupRepository.findById(jointShoppingGroupNum)
-                .orElseThrow(() -> new NotFoundException("해당 번호에 맞는 모임이 없습니다. code : " + jointShoppingGroupNum));
-
-        return jointShoppingGroup.getJointShoppingGroupMaximumCount();
-    }
-
-    /* 최대 참가자 수를 반환하는 로직 */
-    public int findParticipationMaximumCount(Long jointShoppingGroupNum) {
-
-        JointShoppingGroup jointShoppingGroup = jointShoppingGroupRepository.findById(jointShoppingGroupNum)
-                .orElseThrow(() -> new NotFoundException("해당 번호에 맞는 모임이 없습니다. code : " + jointShoppingGroupNum));
-
-        return jointShoppingGroup.getJointShoppingParticipationMaximumCount();
     }
 
     /* 인원수가 가득 찼을 때 마감 상태로 변경하는 로직 */
@@ -124,18 +115,20 @@ public class JointShoppingGroupDomainService {
     }
 
     /* 나가거나 강퇴됬을 때 신청가능 상태로 변경하는 로직 */
-    public void updateApplication(Long jointShoppingGroupNum) {
-        JointShoppingGroup jointShoppingGroup = jointShoppingGroupRepository.findById(jointShoppingGroupNum)
-                .orElseThrow(() -> new NotFoundException("해당 번호에 맞는 모임이 없습니다. code : " + jointShoppingGroupNum));
-
+    public void updateApplication(JointShoppingGroup jointShoppingGroup) {
         /* 수정을 위해 엔터티 정보 변경 */
         jointShoppingGroup.changeApplication();
     }
 
     /* 배송 정보를 수정하는 로직 */
-    public void updateDeliveryInfo(Long jointShoppingGroupNum, JointShoppingDeliveryInfoRequest jointShoppingDeliveryInfoRequest) {
+    public void updateDeliveryInfo(Long jointShoppingGroupNum, JointShoppingDeliveryInfoRequest jointShoppingDeliveryInfoRequest, User user) {
+
         JointShoppingGroup jointShoppingGroup = jointShoppingGroupRepository.findById(jointShoppingGroupNum)
                 .orElseThrow(() -> new NotFoundException("해당 번호에 맞는 모임이 없습니다. code : " + jointShoppingGroupNum));
+
+        if(!user.getUserId().equals(jointShoppingGroup.getUser().getUserId()) ){
+            throw new BadJoinException("방장이 아니여서 배송 정보를 입력하실 수 없습니다.");
+        }
 
         /* 수정을 위해 엔터티 정보 변경 */
         jointShoppingGroup.updateDeliveryInfo(
@@ -145,12 +138,9 @@ public class JointShoppingGroupDomainService {
     }
 
     /* 상품 상태를 체크하고 모집중 상태가 아니라면 취소가 불가능하게 하는 로직 */
-    public void checkProductsStateRecruitment(Long jointShoppingGroupNum) {
-        JointShoppingGroup jointShoppingGroup = jointShoppingGroupRepository.findById(jointShoppingGroupNum)
-                .orElseThrow(() -> new NotFoundException("해당 번호에 맞는 모임이 없습니다. code : " + jointShoppingGroupNum));
-
+    public void checkProductsStateRecruitment(JointShoppingGroup jointShoppingGroup) {
         String state = jointShoppingGroup.getJointShoppingProductsState().toString();
-        if(!state.equals("Recruitment")){
+        if (!state.equals("Recruitment")) {
             throw new BadJoinException("이미 배송이 시작되어 취소할 수 없습니다.");
         }
     }
@@ -161,16 +151,13 @@ public class JointShoppingGroupDomainService {
                 .orElseThrow(() -> new NotFoundException("해당 번호에 맞는 모임이 없습니다. code : " + jointShoppingGroupNum));
 
         String state = jointShoppingGroup.getJointShoppingProductsState().toString();
-        if(!state.equals("OrderCompleted")){
+        if (!state.equals("OrderCompleted")) {
             throw new BadJoinException("아직 배송 정보를 입력하실 수 없습니다.");
         }
     }
 
     /* 상품 상태를 변경하는 로직 */
-    public void changeProductsState(Long jointShoppingGroupNum) {
-        JointShoppingGroup jointShoppingGroup = jointShoppingGroupRepository.findById(jointShoppingGroupNum)
-                .orElseThrow(() -> new NotFoundException("해당 번호에 맞는 모임이 없습니다. code : " + jointShoppingGroupNum));
-
+    public void changeProductsState(JointShoppingGroup jointShoppingGroup) {
         /* 수정을 위해 엔터티 정보 변경 */
         jointShoppingGroup.changeProductsState(
                 jointShoppingGroup.getJointShoppingProductsState()
